Refactor Node service to implement Memes API
- Updated package.json to include author and change Node engine version.
- Refactored MemeController to handle new meme attributes (title, imageUrl, tags, likes).
- Modified database queries in MemeController to reflect new schema.
- Removed excursions.js route file and consolidated routes into memes.js.
- Updated index.js to reflect new API structure and endpoints.
- Changed Meme model to accommodate new attributes and validation rules.
- Updated tests to cover new meme functionality and ensure proper validation.
- Adjusted setup-test-db.js and related test files to create and seed memes table.

+++ software_templates/node-service/skeleton/eslint.config.js
@@ -28,7 +28,7 @@ export default [
       eqeqeq: ["error", "always"],
       curly: ["error", "all"],
       semi: ["error", "always"],
-      quotes: ["error", "single", { avoidEscape: true }],
+      quotes: ["error", "double", { avoidEscape: true }],
     },
   },
   {
